FUK THIS SHIT

+++ API/Modals/TideModal.java
@@ -8,17 +8,17 @@ public class TideModal {
      * DayLow: datetime
      * DayHigh: datetime
      * Going: BIT
-     * Direction: float
+     * Direction: Double
      * FK Name: varchar(50)
      *****************************************/
     public String TS;
     public String DayLow;
     public String DayHigh;
     public boolean Going;
-    public float Direction;
+    public Double Direction;
     public String Name;
 
-    public TideModal(String TS, String DayLow, String DayHigh, boolean Going, float Direction, String Name) {
+    public TideModal(String TS, String DayLow, String DayHigh, boolean Going, Double Direction, String Name) {
         this.TS = TS;
         this.DayLow = DayLow;
         this.DayHigh = DayHigh;
