+++ API/Modals/WindModal.java
@@ -7,16 +7,16 @@ public class WindModal {
      * Create_WindTable.sql
      * Scheme:
      * PK TS: datetime
-     * Speed: float
-     * Direction: float
+     * Speed: Double
+     * Direction: Double
      * FK Name: varchar(50)
      *******************************************/
     public Date TS;
-    public float Speed;
-    public float Direction;
+    public Double Speed;
+    public Double Direction;
     public String Name;
 
-    public WindModal(Date TS, float Speed, float Direction, String Name) {
+    public WindModal(Date TS, Double Speed, Double Direction, String Name) {
         this.TS = TS;
         this.Speed = Speed;
         this.Direction = Direction;
